A lot of small changes, mostly fixing small inconsistencies

+++ src/main/java/com/kuborros/FurBotNeo/commands/MusicCommands/MusicTimeCommand.java
@@ -40,7 +40,7 @@ public class MusicTimeCommand extends MusicCommand {
         int seconds;
         try {
             seconds = (min ? 60 : 1) * Integer.parseInt(val);
-            Long milis = (long) (seconds * 1000);
+            long milis = (seconds * 1000);
             if (getPlayer(guild).getPlayingTrack().getDuration() <= milis) {
                 event.replyWarning("This track is not long enough to skip that far!");
                 return;
